GUACAMOLE-362: Move password decryption logic into TicketValidationService class.
